Move UID to separate file, refine run sh
